refactor: dedupe CSS assembly and string-shape walking (simplify gate)
- assemble_css factors the theme/utility block assembly shared by
  scope_css and shadow_css_for_component; the two formats differ only in
  selectors.
- class_template_value moves to ts_helpers and extract_string_value is
  now the static-only view of it, removing the parallel string-extraction
  walker.
- match_rules drops a needless block-expression wrapper.
